Decided on another design change
After thinking about it, I decided that using GBoxed wrapped objects as
type readers is nice as an exercise, but redundant for constant objects.

Started to move all type readers to functions, created a system to
register new reader functions, and started working on a system to
support parsing of different GBoxed and GObject types.